Re-use root validator for options
