+++ vizro-core/src/vizro/models/_components/form/checklist.py
@@ -1,11 +1,11 @@
 from typing import List, Literal, Optional
 
 from dash import dcc, html
-from pydantic import Field
+from pydantic import Field, root_validator
 
 from vizro.models import Action, VizroBaseModel
 from vizro.models._action._actions_chain import _action_validator_factory
-from vizro.models._components.form._form_utils import get_options_and_default
+from vizro.models._components.form._form_utils import get_options_and_default, validate_options_dict
 from vizro.models._models_utils import _log_call
 from vizro.models.types import MultiValueType, OptionsType
 
@@ -22,17 +22,15 @@ class Checklist(VizroBaseModel):
     """
 
     type: Literal["checklist"] = "checklist"
-    options: OptionsType = Field([])
-    value: Optional[MultiValueType] = Field(None)
+    options: OptionsType = []  # type: ignore[assignment]
+    value: Optional[MultiValueType] = None
     title: Optional[str] = Field(None, description="Title to be displayed")
     actions: List[Action] = []
 
-    # validator
-    set_actions = _action_validator_factory("value")  # type: ignore[pydantic-field]
+    # Validators
+    _set_actions = _action_validator_factory("value")  # type: ignore[pydantic-field]
+    _validate_options_dict = root_validator(allow_reuse=True, pre=True)(validate_options_dict)
 
-    # This should be put back if we expose multi at top level of Filter.
-    # It will ensure there's an explicit error message  if someone calls with multi=False.
-    # multi: Literal[True] = True
     @_log_call
     def build(self):
         full_options, default_value = get_options_and_default(options=self.options, multi=True)
